feat: edit syncStorages func

+++ lib/features/tasks/data/repository/tasks_repository.dart
@@ -1,4 +1,5 @@
-import 'package:todo_list_app/core/utils/managers/persistence_manager.dart';
+import 'package:todo_list_app/core/managers/network_manager.dart';
+import 'package:todo_list_app/core/managers/persistence_manager.dart';
 import 'package:todo_list_app/features/tasks/data/api/network_storage_tasks_api.dart';
 import 'package:todo_list_app/features/tasks/data/api/tasks_api.dart';
 import 'package:todo_list_app/features/tasks/domain/response_data.dart';
@@ -6,8 +7,8 @@ import 'package:todo_list_app/features/tasks/domain/task_model.dart';
 
 class TasksRepository {
   TasksRepository({
-    required networkManager,
-    required persistenceManager,
+    required NetworkManager networkManager,
+    required PersistenceManager persistenceManager,
     required this.localStorage,
   }) : _persistenceManager = persistenceManager {
     networkStorage = NetworkStorageTasksApi(
@@ -26,28 +27,31 @@ class TasksRepository {
   }
 
   Future<void> syncStorages() async {
-    final networkRevision = (await networkStorage.getTasks()).revision;
-    if (await _persistenceManager.getTasksRevision() != networkRevision ||
+    final network = await networkStorage.getTasks();
+    if (await _persistenceManager.getTasksRevision() != network.revision ||
         await checkChanges()) {
       await _persistenceManager.saveTasksRevision(
-          revision: networkRevision ?? 0);
+          revision: network.revision ?? 0);
       final localTasks = await getLocalTasks();
-      final syncTasksList = (await networkStorage.syncTasks(localTasks)).data;
-      if (syncTasksList != null) {
-        for (final task in syncTasksList) {
-          if (localTasks
-              .where((localTask) => localTask.id == task.id)
-              .isEmpty) {
+
+      if (network.data != null) {
+        final localTasksMap = <String, Task>{
+          for (var task in localTasks) task.id: task
+        };
+        for (final task in network.data!) {
+          if (!localTasksMap.containsKey(task.id)) {
             await localStorage.addTask(task);
           } else {
-            final tempTask =
-                localTasks.firstWhere((localTask) => localTask.id == task.id);
-            if (tempTask.changedAt.isBefore(task.changedAt)) {
+            final tempTask = localTasksMap[task.id];
+            if (tempTask!.changedAt.isBefore(task.changedAt)) {
               await localStorage.updateTask(task);
+            } else if (tempTask.deleted != null && tempTask.deleted!) {
+              await localStorage.deleteTask(tempTask.id);
             }
           }
         }
       }
+      await networkStorage.syncTasks(await localStorage.getTasks());
     }
   }
 
@@ -70,7 +74,7 @@ class TasksRepository {
   Future<Task> addTask(Task task) async {
     final localTask = await localStorage.addTask(task);
     final ResponseData response = await networkStorage.addTask(task);
-    if (response.revision != await _persistenceManager.getTasksRevision()) {
+    if (response.status == 400) {
       await syncStorages();
       await networkStorage.addTask(task);
     }
@@ -80,7 +84,7 @@ class TasksRepository {
   Future<Task> updateTask(Task task) async {
     final localTask = await localStorage.updateTask(task);
     final ResponseData response = await networkStorage.updateTask(task);
-    if (response.revision != await _persistenceManager.getTasksRevision()) {
+    if (response.status == 400) {
       await syncStorages();
       await networkStorage.updateTask(task);
     }
@@ -88,11 +92,16 @@ class TasksRepository {
   }
 
   Future<void> deleteTask(String id) async {
-    await localStorage.deleteTask(id);
+    await localStorage.deleteTaskWithoutInternet(id);
     final ResponseData response = await networkStorage.deleteTask(id);
-    if (response.revision != await _persistenceManager.getTasksRevision()) {
+    ResponseData? responseAfterSync;
+    if (response.status == 400) {
       await syncStorages();
-      await networkStorage.deleteTask(id);
+      responseAfterSync = await networkStorage.deleteTask(id);
+    }
+    if (response.status == 200 ||
+        (responseAfterSync != null && responseAfterSync.status == 200)) {
+      await localStorage.deleteTask(id);
     }
   }
 }
